Extract reply_retry.rs from turn.rs (parent #680)
Eighth vertical-slice extraction: move the assistant-reply retry
orchestration out of turn.rs into a new
src/agent/loop_run/reply_retry.rs sibling module. Largest free-fn
conversion in the series (11 methods).

Migrated (all free fns over &mut Agent / &Agent):

Entry point (pub(super)):
- request_assistant_reply_with_retry — freezes footer, starts spinner,
  loops over request_assistant_reply with branch-specific recovery.

Error-routing core + branch handlers (private):
- handle_assistant_reply_retry_error
- maybe_disable_native_tools_after_request_error
- maybe_handle_assistant_reply_format_error
- push_tool_call_format_retry_note
- maybe_handle_assistant_reply_timeout_error
- maybe_handle_assistant_reply_transport_error
- finish_assistant_reply_retry

Ollama dispatch (private):
- request_assistant_reply
- request_streaming_assistant_reply

Format-error terminator (private):
- maybe_finish_after_edit_format_error (Issue #634)

`current_assistant_model` stays on `impl Agent` (5+ external call
sites: turn.rs ×2, scaffold_pipeline.rs ×2, actor_loop_flow.rs ×1).
The new module reaches it via `agent.current_assistant_model()`.

Visibility promotions in turn.rs:
- `AssistantReplyRetryState` struct + 7 fields → pub(super)
- `AssistantReplyRetryDecision` enum → pub(super)
- `AssistantReplyRetryState::new` → pub(super)
- `build_request_messages` method → pub(super)
- `tool_specs_for_policy` method → pub(super)

No facade re-export (DR3-001). Caller updates:
- actor_loop_flow.rs: `agent.request_assistant_reply_with_retry(...)` →
  `super::reply_retry::request_assistant_reply_with_retry(agent, ...)`

`cargo fix --tests` pruned 5 newly-unused imports from turn.rs's prelude
+ 1 from loop_run.rs + 1 from the new module.

turn.rs: 5,736 -> 5,368 (-368 LOC)
Cumulative from 39,489: -34,121 LOC (-86.4%)

Verification:
- cargo fmt --check
- cargo clippy --lib --all-targets -- -D warnings
- cargo test --lib (3,114 passed)
